Make antagonist intervals generate the same layer label as thread intervals.
Antagonist intervals did not override interval's label method, which caused it to use the default one. Since thread intervals do override said method, and since layers are deduped by their label, toggling on an layer  in the antagonist table did not link the layer to the appropriate row in the thread table.

PiperOrigin-RevId: 261159994