Avoiding racy double-creation of Collections when multiple requests on the same collection are sent by the FE to the BE at once.
This is accomplished by equipping each CachedCollection with a 'ready' channel that is closed only once the CachedCollection will no longer be modified.  When GetCollection implementations must add their collection to the cache, they first place an empty CachedCollection in the cache (via StorageService.getCollectionFromCache, which now has an extra boolean argument signifying whether the collection should be added to the cache if it is not yet present there.)  Then, they populate that CachedCollection, calling release() on it when it is fully populated (or cannot be released due to an error.)

When GetCollection implementations find the sought collection already in the cache, they call wait() on it, so that it blocks until it has been released.

This ensures that, even if the FE sends multiple simultaneous requests for the same collection to the backend, it will be processed only once.

This CL also simplifies EditCollection: the cached collection is now simply removed and reloaded, rather than changed-in-place, which has significant synchronization challenges.

PiperOrigin-RevId: 272758130