Collector and some loading logic for sched traces gathered by eBPF.
Includes a bpftrace script, a shell script to invoke it, and an importer library that converts the script's output to an EventSet viewable by SchedViz.

While checking out imported eBPF traces, I found that LoadSchedSwitchWithSynthetics had an undesirable behavior: it assumed switching-in tasks' prior state was Waiting, but switching-in from Sleeping is not uncommon, and assuming Waiting when the thread was actually Sleeping tends to yield spurious front-heavy Waiting intervals.  Unknown is certainly more accurate, so I've switched to that.

PiperOrigin-RevId: 276306574