Workaround for non-contiguous core IDs in topology.
Some machines may not report their core IDs in a contiguous manner. SchedViz assumes that core IDs are always contiguous when drawing the topology. This change uses the maximum core ID instead of the block size as the upper bound of the contiguous range of cores. This wastes some memory (for every "hole" in the core ID list), but works correctly. I also changed the sorting algorithm for the topology list to one that returns the same ordering, but isn't dependent on block size.

PiperOrigin-RevId: 323851879