Fix bug where collection metadata's CPU list would be used instead of the system topology CPU list.
SchedViz is supposed to use the system topology list as the source of truth and fall back to the metadata if the machine has no topology. In this case, the collection metadata was always being used due to a bad usage (or undefined behavior?) of TypeScript parameter properties.
The offending constructor declared a field using parameter properties, called its super constructor which declared and set that same field, and then used the value of the field. The values passed to the sub and super class were not the same, and the generated code set the sub class's value of the field to the subclass constructor argument instead of to the superclass's (as I expected it would as user code written in the super class constructor would do so). Removing the visibility modifier fixes this problem by making it only a constructor argument and it doesn't change anything else as the subclass inherits field from its superclass.

PiperOrigin-RevId: 357021189