Fixing a bug in ThreadSummaries query.
In that query, in order to capture migrations (which may be to or from filtered-out CPUs), we override the overall query's CPU filter when fetching thread intervals, then reapply CPU filters when assembling the summary Metrics.

By convention, when declaring a filter, an empty set of CPUs filtered-in (`CPUs()`) or an unspecified CPU filter is interpreted as `all CPUs`.  However, within a filter f, the f.cpu field has been corrected to reflect the actual set of CPUs filtered in -- if empty, no CPUs were filtered in.

Previously, we were using f.cpu to create a new CPU filter.  When no CPUs were actually filtered in (in practice, this can occur when the topology mentions CPUs that aren't attested in the trace, for instance when those CPUs' scheduling behavior was static over the entire trace and then the user filtered down to only such CPUs), we would then create a new filter for metrics construction with an empty CPU set, which is then expanded to 'all CPUs'.

PiperOrigin-RevId: 360965011